Upgrade Celery and friends to latest versions
We are receiving "Acquire on closed pool" error randomly after running instances
for more than ~1 day and calling `self.app.control.cancel_consumer` from our
task that kills VM instances.

This seems to be a known problem in 3.x Celery verions but some users reported
that it's still present in <4.4.

- https://github.com/celery/celery/issues/4410#issuecomment-505728049
- https://stackoverflow.com/questions/36789805/celery-kombu-fails-after-self-connections-acquire
- https://github.com/celery/celery/issues/1839

Celery released 5.x on September, so I'm upgrading to it directly as a test. If
everything keep working together, we can leave it. Otherwise, we can go back to
latest 4.4.x series: 4.4.7.