Build: don't track changed files
We aren't caching per file,
and to do this we are tracking the md5 for each file.
Which involves reading te file, calculate the md5, and query the db for
each html file.

We also don't need the HTMLFile model at all..
But, we can remove that in another PR,
this should help to make indexing fast.